debugging htree

+++ a11/htree.c
@@ -81,7 +81,7 @@ void *hash( void *ptr )
     if(left < m) pthread_create(&thread1, NULL, hash, (void *) (2 * i + 1));
     if(right < m) pthread_create(&thread2, NULL, hash, (void *) (2 * i + 2));
 
-    size_t bytesPerThread = blocksPerThread * BLOCK_SIZE;
+    long int bytesPerThread = blocksPerThread * BLOCK_SIZE;
     // calc assigned hash (i --> from i * n / m to i * n / m + n / m)
     uint8_t key[bytesPerThread];
     
@@ -91,7 +91,7 @@ void *hash( void *ptr )
     
     printf("i: %d\n", i);
 
-    for(size_t j = 0; j < bytesPerThread; j++)
+    for(long int j = 0; j < bytesPerThread; j++)
         printf("%c", key[j]);
     printf("\n");
